WIP for tiles at different heights. Game looks like ass

+++ centauri-tac/Assets/Scripts/game/models/MapImportModel.cs
@@ -17,7 +17,7 @@ namespace ctac
     public class TileImportPosition
     {
         public int x { get; set; }
-        public int y { get; set; }
+        public float y { get; set; }
         public int z { get; set; }
     }
 }
